image-inpainting/requirements.txt
gradio>=4.0.0
simple-lama-inpainting>=0.1.2
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
